chore(venue): refactor into different files
